DoseCalculations.AbstractBeamLimitingDevice — Type

Abstract type for Beam Limiting Devices are based on

AbstractBeamLimitingDevice contains the types for MultiLeafCollimators, Jaws, etc.

DoseCalculations.AbstractBixel — Type

AbstractBixel{T<:AbstractFloat}

DoseCalculations.AbstractExternalSurface — Type

AbstractExternalSurface

Supertype for external surfaces.

DoseCalculations.AbstractTreatmentField — Type

Treatment Field

Abstract treatment field, basis for containing multiple treatment types (e.g. VMAT, IMRT, etc.)

DoseCalculations.Bixel — Type

Bixel{T}

DoseCalculations.ConstantSurface — Type

ConstantSurface

An external surface with a constant source surface distance, regardless of the position. This surface is largely used for testing purposes as its inherently unphysical.

DoseCalculations.IsoplaneSurface — Type

IsoplaneSurface

An external surface collapsed onto the iso-plane.

Can be precomputed for given gantry angles, avoiding costly mesh-segment intersections.

DoseCalculations.IsoplaneSurface — Method

IsoplaneSurface(xg, yg, SAD)

Construct a IsoplaneSurface out of two axes and the Source-Axis Distance (SAD).

DoseCalculations.IsoplaneSurface — Method

IsoplaneSurface(xg, yg, SAD)

Construct a IsoplaneSurface out of two axes and the Source-Axis Distance (SAD).

DoseCalculations.Jaws — Type

Jaws

Jaws stores the x and y positions of the jaws.

The x/y positions of the jaws can be accessed through the getx/gety methods.

The usual constructor directly takes the jaw x and y position vectors, but a single fieldsize can also be specified.

DoseCalculations.Jaws — Method

Jaws(fieldsize::T)

Create jaws with a square field of length fieldsize.

DoseCalculations.MeshSurface — Type

MeshSurface

An external surface defined by a 3D mesh.

DoseCalculations.MultiLeafCollimator — Type

MultiLeafCollimator

MultiLeafCollimator stores the leaf positions and edges of an MLC.

The 2 by n leaf positions are stored in the positions matrix. The leaf position boundaries are stored in the edges vector, which is 1 element longer than the number of leaf tracks, n.

Indexing a MultiLeafCollimator will either return the x and y bounds of the leaf track, or another MultiLeafCollimator. It also supports the @view macro. Iteration goes through each leaf track, returning the lower and upper boundaries.

A MultiLeafCollimator can also be shifted through addition or subtraction, and scaled with multplication and division.

The locate method is implemented to return the leaf index containing edge position.

DoseCalculations.MultiLeafCollimator — Method

MultiLeafCollimator(n::Int, Δy::Real)

Construct an MLC with leaf edges.

DoseCalculations.MultiLeafCollimator — Method

MultiLeafCollimator(n::Int, Δy::Real)

Construct an MLC with n number of leaves and of leaf width Δy, centered about zero.

DoseCalculations.MultiLeafCollimatorSequence — Type

MultiLeafCollimatorSequence

MultiLeafCollimatorSequence stores a sequence of MLC apertures.

The 2 by n leaf positions are stored in the positions matrix. The leaf position boundaries are stored in the edges vector, which is 1 element longer than the number of leaf tracks, n.

Indexing a MultiLeafCollimator will either return the x and y bounds of the leaf track, or another MultiLeafCollimator. It also supports the @view macro. Iteration goes through each leaf track, returning the lower and upper boundaries.

A MultiLeafCollimator can also be shifted through addition or subtraction, and scaled with multplication and division.

The locate method is implemented to return the leaf index containing edge position.

DoseCalculations.PlaneSurface — Type

PlaneSurface

A planar external surface at a constant distance from the source.

Base.getindex — Method

Base.getindex(bixel::AbstractBixel, i::Int)

Get the centre position of the bixel.

Base.iterate — Method

iterate

Iteration of a treatment field, returning ControlPoint every time iterate is called.

DoseCalculations.area — Method

area(bixel::AbstractBixel)

Return the area of the bixel.

DoseCalculations.axes_lims! — Method

axes_lims!(p, jaws::Jaws; pad=10.)

Set the axes limits to the position of the jaws.

By default, a pad of 10 mm is added to each side.

DoseCalculations.axes_lims! — Method

axes_lims!(jaws::Jaws; pad=10.)

When no plot object given, applies to latest plot created.

DoseCalculations.bixel_grid — Method

bixel_grid(x, y, Δx[, Δy])

Construct a grid of bixels.

Each axis starts at the first element (e.g. x[1]), runs to the last element (x[end]), with uniform spacing (Δx). Same for y. Positions are "snapped" to the spacing value (see snapped_range for details).

If Δy not specified, assumes Δy = Δx.

DoseCalculations.bixel_grid — Method

bixel_grid(x::AbstractRange, y::AbstractRange)

Uses the start and end positions and step of each range to construct the bixel grid.

DoseCalculations.bixel_grid — Method

bixel_grid(mlc::AbstractMultiLeafCollimator, jaws::Jaws, Δx)

Grid that fits in an MLC and the jaws.

Bixel y widths are of the same width as the MLC leaf widths. Creates smaller widths in the case where the jaws are halfway within a leaf width. Bixel x widths are set by Δx.`

DoseCalculations.bixel_grid — Method

bixel_grid(jaws::Jaws, Δx[, Δy])

Uses the jaw positions to construct a bixel grid.

If Δy not specified, assumes Δy = Δx

DoseCalculations.bixels_from_bld — Function

bixels_from_bld(args::AbstractBeamLimitingDevice...)

Create bixels corresponding to the provided beam limiting devices.

DoseCalculations.bixels_from_bld — Method

bixel_from_bld(jaws::Jaws)

From Jaws.

DoseCalculations.bixels_from_bld — Method

bixels_from_bld(mlcx, mlc::MultiLeafCollimator, jaws::Jaws)

From MultiLeafCollimator and Jaws.

DoseCalculations.bld_to_fixed — Method

bld_to_fixed(ϕg, θb, SAD)

Convert from IEC BLD to IEC fixed coordinates

DoseCalculations.bld_to_gantry — Method

bld_to_gantry(θb, SAD)

Convert from IEC BLD to IEC gantry coordinates

DoseCalculations.compute_SSD! — Method

compute_SSD!(surf::IsoplaneSurface, mesh_surface::MeshSurface)

Compute the SSD matrix from the 3D mesh stored in MeshSurface.

DoseCalculations.dose_fluence_matrix! — Method

dose_fluence_matrix!(D, pos, bixels, bixels, surf, calc)

Compute a dose-fluence matrix from dose positions, bixels, external surface and dose calculation algorithm.

Requires the point_kernel! method to be defined for the given dose calculation algorithm (calc). point_kernel! computes the dose calculated from the set of bixels a given dose point. Stores result in D.

DoseCalculations.dose_fluence_matrix — Method

dose_fluence_matrix(pos, bixels, bixels, surf, calc)

Compute a dose-fluence matrix from dose positions, bixels, external surface and dose calculation algorithm.

See dose_fluence_matrix! for implementation.

DoseCalculations.dose_from_beam — Method

dose_from_beam(pos, surf, beam, calc)

Compute the dose from a beam.

DoseCalculations.eclipse_to_SIK — Function

eclipse_to_SIK(input_datadir)

Read in Eclipse exported files, extracting Scaled Isoplane Kernel data

Requires SinglePencilBeamKernel{depth} for multiple depths and MeasuredProfile files in input_datadir. Returns a dict with kernel and depth dose data.

DoseCalculations.eclipse_to_SPBK — Method

eclipse_to_SPBK(input_datadir)

Read in Eclipse exported files, extracting Single Pencil Beam Kernel data

Requires SinglePencilBeamKernel{depth} and IntensityProf{depth} files in input_datadir for multiple depths. Returns a dict with kernel and intensity profile data.

DoseCalculations.fixed_to_bld — Method

fixed_to_bld(ϕg, θb, SAD)

Convert from IEC fixed to IEC BLD coordinates

DoseCalculations.fixed_to_gantry — Method

fixed_to_gantry(ϕg)

Convert from IEC fixed to IEC gantry coordinates

DoseCalculations.fluence — Method

fluence(bixels::AbstractArray{<:AbstractBixel}, index::AbstractArray{Int}, args...)

Allows precomputation of the location of the bixel in relation to the beam limiting device.

In the case of an MLC, the index is the leaf index that contains that bixel.

Requires fluence(bixel::AbstractBixel, index::Int, args...) to be defined for the particular beam limiting device

DoseCalculations.fluence — Method

fluence(bixels::AbstractArray{<:AbstractBixel}, bld::AbstractBeamLimitingDevice, args...)

Compute the fluence on a collection of bixels.

Broadcasts over the specific fluence(bixel, ...) method for the provided beam limiting device. e.g.: fluence(bixels, jaws), fluence(bixels, mlcx, mlc)

DoseCalculations.fluence — Method

fluence(bixel::Bixel, jaws::Jaws)

From the Jaws.

DoseCalculations.fluence — Method

fluence(bixel::AbstractBixel, bld::AbstractBeamLimitingDevice, args...)

Compute the fluence of bixel from beam limiting device (e.g. an MLC or jaws).

DoseCalculations.fluence — Method

fluence(bixel::AbstractBixel{T}, index::Int, mlcx1, mlcx2)

From an MLC aperture sequence using a given leaf index.

DoseCalculations.fluence — Method

fluence(bixel::Bixel, index::Int, mlcx)

From an MLC aperture using a given leaf index.

This method assumes the bixel is entirely within the ith leaf track, and does not overlap with other leaves. Does not check whether these assumptions are true.

DoseCalculations.fluence — Method

fluence(bixel::AbstractBixel{T}, mlc1::MultiLeafCollimator, mlc2::MultiLeafCollimator)

From an MLC aperture sequence.

DoseCalculations.fluence — Method

fluence(bixel::Bixel, mlc::MultiLeafCollimator)

From an MLC aperture.

DoseCalculations.fluence_from_moving_aperture — Method

fluence_from_moving_aperture(bixel::AbstractBixel{T}, mlcx1, mlcx2)

From MLC leaf positions which move from mlcx1 to mlcx2.

Computes the time-weighted fluence as the MLC moves from position mlcx1 to mlcx2. Assumes the MLC leaves move in a straight line.

DoseCalculations.fluence_from_rectangle — Method

fluence_from_rectangle(bixel::Bixel, xlim, ylim)

Compute the fluence of a rectangle with edges at xlim and ylim on a bixel.

DoseCalculations.fluence_grid_length — Method

fluence_grid_length(calc::AbstractDoseAlgorithm)

Return number of bixels in the fluence grid.

DoseCalculations.fluence_grid_size — Method

fluence_grid_size(calc::AbstractDoseAlgorithm)

Return the size of the fluence grid.

DoseCalculations.fluence_onesided — Method

fluence_onesided(xs, xf, xL, xU)

Compute fluence for 1 leaf position, assuming the other is infinitely far away

Computes the fluence for a leaf trajectory from xs to xf, over a bixel from xL to xU. The aperture is considered open to the right of the leaf position (i.e. leaf on the B bank). Assumes the bixel is fully in the leaf track.

DoseCalculations.gantry_to_bld — Method

gantry_to_bld(θb, SAD)

Convert from IEC gantry to IEC BLD coordinates

DoseCalculations.gantry_to_fixed — Method

gantry_to_fixed(ϕg)

Convert from IEC gantry to IEC fixed coordinates

DoseCalculations.getSSD — Method

getSSD(calc::ConstantSurface, pᵢ)

When applied to a ConstantSurface, it returns a constant SSD value, regardless of position pᵢ.

DoseCalculations.getSSD — Method

getSSD(surf::AbstractExternalSurface, pᵢ)

Get the Source-Surface Distance (SSD) between for the given position pᵢ.

The position pᵢ is in the IEC Beam Limiting Device (BLD) coordinate system (whose center is at the radiation source).

DoseCalculations.getSSD — Method

getSSD(surf::IsoplaneSurface, pᵢ)

When applied to a IsoplaneSurface, it locates the position on the isoplane then returns the source-surface distance.

Must call compute_SSD! on surf first.

DoseCalculations.getSSD — Method

getSSD(surf::PlaneSurface, pᵢ)

When applied to a PlaneSurface, it returns the distance to the plane.

It assumes the plane is at position (0, 0, surf.sourcesurfacedistance) with normal pointing towards the source

DoseCalculations.getSSD — Method

getSSD(surf::MeshSurface, pᵢ)

When applied to a MeshSurface, it returns the smallest distance to the mesh.

DoseCalculations.get_dataset_line_numbers — Method

getdatasetline_numbers(lines)

Get the line numbers of each data section.

Looks for the string "Profiles" in each line, then adds the line number to the list of dataset line numbers, which is returned. See readeclipseprofile_data

DoseCalculations.getdepth — Method

getdepth(surf::AbstractExternalSurface, pᵢ)

Get the depth of the position pᵢ below the surface surf.

DoseCalculations.integrate_kernel — Method

integrate_kernel(calc::ScaledIsoplaneKernel, bixel::AbstractBixel{T}, x_iso, y_iso) where T<:AbstractFloat

Integrate the kernel over bixel from position x_iso, y_iso.

Subdivides the bixel for higher accuracy.

DoseCalculations.interp — Method

interp(xg::AbstractVector, yg::AbstractVector, fg::AbstractMatrix, xi, yi)

Bilinear interpolation at position xi,yi, on grid xg-yg with values fg

DoseCalculations.interp — Method

interp(f1, f2, α)

Interpolate using given values f1 and f2 and the scaled distance from them α

DoseCalculations.interp — Method

interp(xg::AbstractVector{T}, fg::AbstractVector{T}, xi) where T<:Real

Interpolate xi within uniform grid positions xg and grid values fg

DoseCalculations.intersect_mesh — Method

intersect_mesh(s, mesh)

Find the intersection points of the line and the mesh.

Returns a list of intersection points, and an empty list if none present.

DoseCalculations.intersect_mesh_multi_threaded — Method

intersect_mesh_multi_threaded(s, mesh)

Multi-threaded version of intersect_mesh.

DoseCalculations.intersect_mesh_single_threaded — Method

intersect_mesh_single_threaded(s, mesh)

Single-threaded version of intersect_mesh.

DoseCalculations.kernel — Method

kernel(calc::ScaledIsoplaneKernel, r)

Get the kernel at radius r.

DoseCalculations.kernel_size — Method

kernel_size(calc::ScaledIsoplaneKernel, pos, bixels)

Compute the number of bixels with the max. radius of a given position.

DoseCalculations.leaf_trajectory — Method

leaf_trajectory(x, x1, x2)

Compute the height of position x between (x1, 0) and (x2, 1)

Used in intersection_area.

DoseCalculations.load_beam — Method

load_beam(beam, total_meterset)

Load a beam from a control point sequence in a DICOM RP file.

DoseCalculations.load_dicom — Method

load_dicom(filename)

Load a DICOM RP file into a Vector{TreatmentField}.

DoseCalculations.load_kernel_data — Method

load_kernel_data(filename)

Load a kernel data file in Eclipse format.

DoseCalculations.load_ref_dose — Method

load_ref_dose(beam)

Load a reference dose, used for calculating the meterset in a control point sequence

DoseCalculations.load_structure_from_ply — Method

load_structure_from_ply(filename)

Load a .ply file into a mesh

Uses MeshIO to load the mesh from the file. Unfortunately, MeshIO loads into GeometryBasics.Mesh, not Meshes.SimpleMesh. The rest of the code converts the mesh into a Meshes.SimpleMesh. From https://github.com/JuliaIO/MeshIO.jl/issues/67#issuecomment-913353708

DoseCalculations.locate — Method

locate(x::T, start::T, step::T) where T<:AbstractFloat

Locate the grid index for position xi

Specify the grid starting position (start) and grid spacing (step).

DoseCalculations.locate — Method

locate(mlc::AbstractMultiLeafCollimator, x::Number)

Locate the track index i which contains the position x

DoseCalculations.locate — Method

locate(xg::AbstractRange{T}, xi::T) where T<:Real

When a range is provided, uniform spacing can be assumed.

DoseCalculations.locate — Method

locate(xg::AbstractVector{T}, xi::T) where T<:Real

Locates the grid index for position x within a grid xg. Does not limit out of bounds indices.

DoseCalculations.minmax — Method

minmax(x, l, u)

Return x if l<=x<=u, l if x<l or u if u<x

DoseCalculations.norm_depth_dose — Method

norm_depth_dose(calc::ScaledIsoplaneKernel, d)

Get the normalised depth dose at depth d.

DoseCalculations.overlap — Method

overlap(x, Δx, xB, xA)

Compute the overlapping length of a bixel spanning x-w/2->x+w/2 and the length between xL and xU, normalised to the length of the bixel. If the Bixel fully within range (xL<=x-w/2 && x+w/2<=xU), return 1. If the bixel is fully outside the range (xU<=x-w/2 || x+w/2<=xL), return 0.

DoseCalculations.parse_value — Method

parse_value(string_value)

Tries to parse a string, returning string if parsing failed

DoseCalculations.patient_to_fixed — Method

patient_to_fixed(isocenter)

Convert from the patient-based coordinate system to IEC Fixed.

Isocenter is in the patient-based coordinate system, as specified in the DICOM RP file.

DoseCalculations.plot_bld! — Method

plot_bld!

Main plot command for beam limiting devices (e.g. jaws, MLCs).

Creates a 2D plot of the given beam limiting device in the beam limiting device (BLD) coordinate system, showing their relevant positions (e.g. aperture positions for MLC)

Use plot_bld to create a new plot object, and plot_bld! to add to an existing one.

DoseCalculations.plot_bld! — Method

plot_bld!(p, jaws::Jaws; kwargs...)

When applied to jaws, plots a box indicating the jaw positions

DoseCalculations.plot_bld! — Method

plot_bld!(p, mlcx, mlc::MultiLeafCollimator; kwargs...)

When applied to MultiLeafCollimator with leaf positions, plot the open aperture.

Fills the area obstructed by leaves, unless invert=true where it fills the open aperture.

DoseCalculations.plot_bld — Method

plot_bld

Main plot command for beam limiting devices (e.g. jaws, MLCs).

See ?plot_bld for more information.

DoseCalculations.point_kernel! — Method

point_kernel!(rowval, nzval, pos::AbstractVector{T}, bixels, surf, calc)

Compute the fluence kernel for a given position.

Designed to be used with a dose-fluence matrix of type SparseMatrixCSC. Stores the row value in rowval, and dose value in nzval.

DoseCalculations.position — Method

position(bixel::AbstractBixel, i::Int)

Get the ith coordinate of the position.

DoseCalculations.position — Method

position(bixel::AbstractBixel)

Get the position of the bixel.

DoseCalculations.read_attr — Method

read_attr(line)

Read attribute name and value

DoseCalculations.read_attributes — Method

read_attributes(filename)

Read attributes from Eclipsed exported measured data file. See readeclipsedepthdose and readeclipse_profile.

DoseCalculations.read_data_line — Method

read_data_line(line; delim=',')

Read a line of data of the form: x1, x2, ... where ',' is the delimiter

DoseCalculations.read_dataset — Method

read_dataset(lines)

Read each line in lines and extract the header row, row index and data.

DoseCalculations.read_eclipse_datafile — Function

read_eclipse_datafile(filename::String, truncate=true)

Read an Eclipse data from file.

Reads Eclipse data from files such as SinglePencilBeamKernel, IntensityProf and EnvelopeProfile files. The files are generally of the form:

```
%attr1_name: attr1_value
%attr2_name: attr2_value
    ...
<x1 y1>
<x2 y2>
    ...
<xn yn>
```

Args:

filename: File path

truncate: Whether to remove zeros, default true (see truncate_zeros)

DoseCalculations.read_eclipse_depth_dose_data — Method

read_eclipse_depth_dose_data(filename)

Read an Eclipse exported depth dose file.

This file is of the form:

```
%attr1_name: attr1_value
%attr2_name: attr2_value
    ...
    ,    col1,    col2, ...,    colm
row1,     x11,     x12, ...,    col1m
row2,     x21,     x22, ...,    col2m
    ...
rown, ...
```

Returns a dictionary with attributes, the column header (as given by column legend), the row index (row legend), and a matrix containing the data.

DoseCalculations.read_eclipse_profile_data — Method

read_eclipse_profile_data(filename)

Read an Eclipse exported profile file.

Similar to readeclipsedepthdosedata, except profile files can have multiple datasets at different depths. This file is of the form:

```
%attr1_name: attr1_value
%attr2_name: attr2_value
    ⁞
Profile at d1
    ,    col1,    col2, ...,    colm
row1,     x11,     x12, ...,    col1m
    ⁞
rown, ...
Profile at d2
    ,    col1,    col2, ...,    colm
row1,     x11,     x12, ...,    col1m
    ⁞
rown, ...
    ⁞
```

Returns a dictionary with attributes and profile data. Profile data is stored in a vector under data["Profiles"].

DoseCalculations.reconstruct_dose — Method

reconstruct_dose(pos, surf, plan, calc; Δx=1., ΔMU=2., show_progess=true)

Reconstruct the dose.

Requires dose positions (pos), an external surface (surf), a set of beams (plan), and the dose calculation algorithm (calc). Optional arguments include:

Δx: Size of each bixel in the fluence grid (defaults to 1.)

ΔMU: Meterset increments (defaults to 2.)

show_progess: Whether to display the progress (defaults to true)

DoseCalculations.resample — Method

resample(field, Δη::T; by=:MU)

Resample at uniform steps Δη, from start to finish.

See resample(field, ηₛ::AbstractVector{T}; by=:MU) for details.

DoseCalculations.resample — Method

resample(field, ηₛ::AbstractVector{T}; by=:time)

Resample a treatment field onto new times or meterset values.

Can resample either by time (by=:time) or MU (by=:MU, default).

DoseCalculations.save — Method

save(filename, mesh::SimpleMesh)

Save a SimpleMesh to a VTK file.

DoseCalculations.scale_to_cell — Method

scale_to_cell(x1, x2, xi)

Scale the position xi within the positions x1 and x2.

DoseCalculations.scale_to_isoplane — Method

scale_to_isoplane(pᵢ, z_plane)

Scale the position pᵢ to the plane at distance z_plane.

DoseCalculations.segment_indices — Method

segment_indices(mlcx)

Segment the MLC aperture into separate contained "openings", returning the leaf index.

The open aperture separates when xB[i] > xA[i+1] or xA[i] < xB[i+1]. This index is used to remove plotting artefacts.

Examples

In this case, the aperture has two "openings", at x=-1.->1. and at x=2.->4. These openings are separated in the y direction, but this information is not required to separate the openings.

julia> mlcx = [-1. -1. 2. 2.;
                1.  1. 4. 4.];
julia> segment_indices(mlcx)
3-element Vector{Int64}:
 1
 3
 5

The first aperture opening runs from index 1 -> 2, and the second opening from index 3 -> 4

DoseCalculations.segment_mlc — Method

segment_mlc(mlcx, mlcy)

Segment the MLC aperture into separate contained "openings", return the x and y positions of the aperture.

Calls segment_indices to get the indices for each section, then segmentx and segmenty to retrieve the positions of the aperture.

DoseCalculations.segmentx — Method

segmentx(xL, xR)

Creates a vector of MLC x positions, used for plotting the position of the aperture.

DoseCalculations.segmenty — Method

segmenty(y)

Creates a vector of MLC y positions, used for plotting the position of the aperture.

DoseCalculations.snapped_range — Method

snapped_range(x1, x2, Δ)

Create a range from x1 to x2 which is "snapped" to the step Δ.

Positions are "snapped" to the step value (e.g. a starting position of x[1]-0.2Δx snaps to x[1]-Δx). The new range always includes the start and end points of the original range

Examples:

0.1:1.:9.4 -> 0:1.:10.

DoseCalculations.split_lines_into_datasets — Method

split_lines_into_datasets(lines, sections)

Split the lines into each dataset. See readeclipseprofile_data

DoseCalculations.subdivide — Method

subdivide(x, Δx, δxmax)

Subdivide at position x with width Δx with a max subvision width of δxmax

Returns the starting position, the subdivision width and the number of subdivisions. Allows for use in a loop, where the center of subdivision i is at: x0 + i*δx from 0:nx-1

DoseCalculations.subdivide — Method

subdivide(bixel::AbstractBixel, nx::Integer, ny::Integer)

Subdivide a bixel by specifing the number of partitions nx and ny.

Returns a grid of bixels.

DoseCalculations.subdivide — Method

subdivide(bixel::AbstractBixel{T}, δx::T, δy::T)

Subdivide by specifing widths δx and δy

DoseCalculations.transform! — Method

transform!(mesh::SimpleMesh, trans)

Apply general transformation trans to mesh, modifying the original mesh.

DoseCalculations.transform — Method

transform(mesh::SimpleMesh, trans)

Apply general transformation trans to mesh`.

DoseCalculations.truncate_zeros — Method

truncate_zeros(x, y)

Removes every element after first zero in y and corresponding x

DoseCalculations.width — Method

width(bixel::AbstractBixel, i::Int)

Return the width of the bixel along axis i.

DoseCalculations.width — Method

width(bixel::AbstractBixel)

Return the width of the bixel.