Dose Reconstruction from a Treatment Plan

Roentgen.jl provides a high-level interface to reconstructing dose, enabling quick and easy dose reconstructions.

reconstruct_dose takes a dose volume, Treatment Plan and a dose calculation algorithm, and computes the dose for all fields in the plan. An example of this is provided in the Dose Reconstruction notebook.

Roentgen.reconstruct_dose — Function

reconstruct_dose(vol::AbstractDoseVolume, plan::AbstractTreatmentPlan,
    calc::AbstractDoseAlgorithm; Δx=5., show_progress=true)

Dose reconstruction from a treatment plan.

Requires a dose volume (vol), a treatment plan (plan), and a dose calculationalgorithm (calc). Optional arguments include:

Δx: Size of each bixel in the fluence grid (defaults to 5.)

show_progress: If true (default), displays the progress